add for

+++ entities/topup.go
@@ -3,6 +3,6 @@ package entities
 type Topup struct {
 	ID                int
 	User_ID           string
-	Nominal           float64
+	Nominal           int
 	Tanggal_Transaksi string
 }
